SUBMARINE-338. Refactor e2e code.
