--- backup.travis.yml
@@ -1,123 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements. See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to You under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License. You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-sudo: required
-
-before_cache:
-  - sudo chown -R travis:travis $HOME/.m2
-
-cache:
-  apt: true
-  directories:
-    - ${HOME}/.m2
-    - submarine-workbench/workbench-web/node
-    - submarine-workbench/workbench-web/node_modules
-    - submarine-workbench/workbench-web-ng/node
-    - submarine-workbench/workbench-web-ng/node_modules
-
-addons:
-  apt:
-    sources:
-      - mysql-5.7-trusty
-    packages:
-      - mysql-server
-      - mysql-client
-
-services:
-  - mysql
-  - xvfb
-
-env:
-  global:
-    # If you need to compile Phadoop-3.1 or Phadoop-3.2, you need to add `!submarine-server/server-submitter/submitter-yarnservice` in EXCLUDE_SUBMARINE
-    - EXCLUDE_SUBMARINE="!submarine-all,!submarine-client,!submarine-commons,!submarine-commons/commons-runtime,!submarine-dist,!submarine-server/server-submitter/submitter-yarn,!submarine-server/server-core,!submarine-server/server-rpc"
-    - EXCLUDE_WORKBENCH="!submarine-workbench,!submarine-workbench/workbench-web,!submarine-workbench/workbench-web-ng"
-    - EXCLUDE_INTERPRETER="!submarine-workbench/interpreter,!submarine-workbench/interpreter/interpreter-engine,!submarine-workbench/interpreter/python-interpreter,!submarine-workbench/interpreter/spark-interpreter"
-    - EXCLUDE_SUBMODULE_TONY="!submodules/tony,!submodules/tony/tony-mini,!submodules/tony/tony-core,!submodules/tony/tony-proxy,!submodules/tony/tony-portal,!submodules/tony/tony-azkaban,!submodules/tony/tony-cli"
-    - EXCLUDE_K8S="!submarine-server/server-submitter/submitter-k8s"
-    - EXCLUDE_COMMON_RPC="!submarine-commons/commons-rpc"
-    - MOZ_HEADLESS=1
-
-before_install:
-  # maven 3.6.1 (3.6.2 build tony failed!!!)
-  - echo "Download Maven 3.6.1"
-  - wget https://archive.apache.org/dist/maven/maven-3/3.6.1/binaries/apache-maven-3.6.1-bin.tar.gz
-  - tar zxvf apache-maven-3.6.1-bin.tar.gz || travis_terminate 1
-  - export M2_HOME=$PWD/apache-maven-3.6.1
-  - export PATH=$M2_HOME/bin:$PATH
-  # mysql
-  - sudo service mysql restart
-  - mysql -e "create database submarine_test;"
-  - mysql -e "CREATE USER 'submarine_test'@'%' IDENTIFIED BY 'password_test';"
-  - mysql -e "GRANT ALL PRIVILEGES ON *.* TO 'submarine_test'@'%';"
-  - mysql -e "use submarine_test; source ./docs/database/submarine.sql; show tables;"
-  - mysql -e "create database metastore_test;"
-  - mysql -e "CREATE USER 'metastore_test'@'%' IDENTIFIED BY 'password_test';"
-  - mysql -e "GRANT ALL PRIVILEGES ON *.* TO 'metastore_test'@'%';"
-  - mysql -e "use metastore_test; source ./docs/database/metastore.sql; show tables;"
-  # For e2e tests
-  - mysql -e "create database submarine;"
-  - mysql -e "CREATE USER 'submarine'@'%' IDENTIFIED BY 'password';"
-  - mysql -e "GRANT ALL PRIVILEGES ON *.* TO 'submarine'@'%';"
-  - mysql -e "use submarine; source ./docs/database/submarine.sql; source ./docs/database/submarine-data.sql; show tables;"
-  - mysql -e "create database metastore;"
-  - mysql -e "CREATE USER 'metastore'@'%' IDENTIFIED BY 'password';"
-  - mysql -e "GRANT ALL PRIVILEGES ON *.* TO 'metastore'@'%';"
-  - mysql -e "use metastore; source ./docs/database/metastore.sql; show tables;"
-  - ./dev-support/travis/install_external_dependencies.sh
-  # protobuf 3.10.1
-  - PROTOBUF_VERSION=3.10.1
-  - PROTOC_FILENAME=protoc-${PROTOBUF_VERSION}-linux-x86_64.zip
-  - pushd /home/travis
-  - wget https://github.com/google/protobuf/releases/download/v${PROTOBUF_VERSION}/${PROTOC_FILENAME}
-  - unzip ${PROTOC_FILENAME}
-  - bin/protoc --version
-  - popd
-
-matrix:
-  include:
-    - name: Test submarine test-e2e
-      language: java
-      jdk: "openjdk8"
-      dist: xenial
-      addons:
-        chrome: stable
-      env: PROFILE="-Phadoop-2.9" BUILD_FLAG="clean package -DskipTests" TEST_FLAG="verify -DskipRat -am" TEST_MODULES="-pl submarine-test/e2e" TEST_PROJECTS=""
-
-install:
-  - mvn --version
-  - echo "mvn $BUILD_FLAG $MODULES $PROFILE -B"
-  - mvn $BUILD_FLAG $MODULES $PROFILE -B
-
-before_script:
-  # https://docs.travis-ci.com/user/gui-and-headless-browsers/#Using-xvfb-to-Run-Tests-That-Require-a-GUI
-  # - if [[ -n $TEST_MODULES ]]; then export DISPLAY=:99.0; echo $DISPLAY; ls -l /etc/init.d/xvfb; sh -e /etc/init.d/xvfb start; sleep 3; fi
-  - echo $DISPLAY
-  - google-chrome -version
-  # display info log for debugging
-  - if [[ -n $TEST_MODULES ]]; then echo "MAVEN_OPTS='-Xms1024M -Xmx2048M -XX:MaxPermSize=1024m -XX:-UseGCOverheadLimit -Dorg.slf4j.simpleLogger.defaultLogLevel=info'" > ~/.mavenrc; fi
-
-script:
-  - if [[ -n $TEST_MODULES ]]; then export MODULES="${TEST_MODULES}"; fi
-  - echo "mvn $TEST_FLAG $MODULES $PROFILE -B $TEST_PROJECTS"
-  - mvn $TEST_FLAG $MODULES $PROFILE -B $TEST_PROJECTS
-
-after_success:
-  - echo "Travis exited with ${TRAVIS_TEST_RESULT}"
-
-after_failure:
-  - echo "Travis exited with ${TRAVIS_TEST_RESULT}"
-  - find . -name rat.txt | xargs cat
-  - cat logs/*
