+++ pom.xml
@@ -131,6 +131,8 @@
     <!--  Submarine on Kubernetes  -->
     <k8s.client-java.version>6.0.1</k8s.client-java.version>
     <jersey.test-framework>2.27</jersey.test-framework>
+    <!-- integration test-->
+    <plugin.failsafe.version>2.17</plugin.failsafe.version>
   </properties>
 
   <modules>
