+++ submarine-test/e2e/src/test/java/org/apache/submarine/WebDriverProvider.java
@@ -1,0 +1,42 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.submarine;
+
+import org.openqa.selenium.WebDriver;
+
+public interface WebDriverProvider {
+  /**
+   * Get webdriver version according to web browser.
+   * @return the required version of web driver
+   */
+  public String getWebDriverVersion();
+
+  /**
+   * Download the web driver.
+   * @param webDriverVersion
+   * @return
+   */
+  public String downloadWebDriver(String webDriverVersion);
+
+  /**
+   * create a new web driver
+   * @param webDriverPath
+   * @return a new web driver
+   */
+  public WebDriver createWebDriver(String webDriverPath);
+}
